+++ packages/plugin-link-preview/web/src/tiptap/component.tsx
@@ -5,7 +5,7 @@ import { PluginProps } from 'wix-rich-content-editor-common';
 import { LINK_PREVIEW_TYPE } from 'ricos-content';
 
 export const LinkPreview: React.FC<PluginProps> = ({ context, componentData }) => {
-  const { isMobile, theme, iframeSandboxDomain, config } = context;
+  const { isMobile, theme, iframeSandboxDomain, config = {} } = context;
   const settings: LinkPreviewPluginEditorConfig = config[LINK_PREVIEW_TYPE] || {};
   const blockProps = {
     setFocusToBlock: () => null,